backend/requirements.txt
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
psycopg2==2.8.6
social-auth-app-django==4.0.0
